extract and refactor
